requirements.txt
pandas
numpy
pyarrow
duckdb
pyYAML

openai
langchain
langgraph
langchain-openai
streamlit

python-dotenv
pydantic
typing-extensions
fastparquet








